Move BMS Table updating to jobs.
